requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic_settings
pydantic[email]
pytest
bcrypt
pyjwt
python-decouple
httpx
faker
alembic
langchain
langchain-openai
langchain_community
langchain_experimental
python-dotenv
APScheduler 